backend/llm/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
pdfplumber>=0.10.0
spacy>=3.0.0
requests>=2.25.0